backend/requirements.txt
fastapi
uvicorn
langchain==0.3.0
langchain-groq
langchain-community==0.3.0
langchain-core==0.3.0
beautifulsoup4
faiss-cpu
langchain-huggingface
python-dotenv
pydantic
requests
lxml
sqlalchemy
psycopg2-binary
pg8000
bcrypt
